mysite/requirements.txt
Django==1.8
django-cors-headers==1.1.0
djangorestframework==3.1.3
